Avoid a middlewares folder and move middlewares to plugins and core
